Bump minSdkVersion to 10 :(
If we use android gradle plugin 3.x, we won't even be able to install the apk on <= eclair devices: https://issuetracker.google.com/issues/111146570
I wasn't able to test this though, as I couldn't get any old emulators to start.

The oldest emulator I could get to start was a gingerbread, so I'm bumping the minSdk to that (10).

Add a notification channel for Oreo+ (or else it's not possible to share a recorded animation.